add lines to invoice endpoint

+++ api/tests/Integration/Invoice/EmitInvoiceEndpointTest.php
@@ -6,7 +6,7 @@ use Test\Integration\EndpointTest;
 
 class EmitInvoiceEndpointTest extends EndpointTest
 {
-    private $invoiceNumber;
+    private string $invoiceNumber = '';
 
     public function test_unauthorized_fails()
     {
@@ -28,8 +28,35 @@ class EmitInvoiceEndpointTest extends EndpointTest
                 'customer_tax_address' => "Cardenal despuig 41",
                 'customer_tax_zip_code' => "07013",
                 'date' => '2023-06-27',
-                'amount' => 1000,
-                'concept' => 'Caja de 12 Moixes',
+                'lines' => [
+                    [
+                        'amount' => 1000,
+                        'concept' => 'Caja de 12 Moixes',
+                        'vat_percent' => 21,
+                    ],
+                ],
+            ], JSON_THROW_ON_ERROR),
+            'headers' => ['Authorization' => 'Bearer ' . $this->authToken],
+        ]);
+
+        $this->assertEquals(400, $response->getStatusCode());
+    }
+
+    public function test_no_lines_fails()
+    {
+        $this->registerUser($this->email, "");
+        $this->login($this->email, "");
+        $this->setBusinessData();
+
+        $response = $this->client->post('/invoice', [
+            'body' => json_encode([
+                'customer_name' => "CUSTOMER Bar",
+                'customer_tax_name' => "Jaume de s'Atomic",
+                'customer_tax_number' => "435678122F",
+                'customer_tax_address' => "Cardenal despuig 41",
+                'customer_tax_zip_code' => "07013",
+                'date' => '2023-06-27',
+                'lines' => [],
             ], JSON_THROW_ON_ERROR),
             'headers' => ['Authorization' => 'Bearer ' . $this->authToken],
         ]);
@@ -51,8 +78,13 @@ class EmitInvoiceEndpointTest extends EndpointTest
                 'customer_tax_address' => "Cardenal despuig 41",
                 'customer_tax_zip_code' => "07013",
                 'date' => '2023-06-27',
-                'amount' => 1000,
-                'concept' => 'Capsa de 12 Moixes',
+                'lines' => [
+                    [
+                        'amount' => 1000,
+                        'concept' => 'Capsa de 12 Moixes',
+                        'vat_percent' => 21,
+                    ],
+                ],
             ], JSON_THROW_ON_ERROR),
             'headers' => ['Authorization' => 'Bearer ' . $this->authToken],
         ]);
@@ -71,8 +103,10 @@ class EmitInvoiceEndpointTest extends EndpointTest
         $this->deleteUser($this->email);
         $this->pdo->query('DELETE FROM business WHERE tax_id = "435678122F";');
         $this->pdo->query('DELETE FROM business WHERE tax_id = "EMIT0012300TEST";');
-        $this->pdo->query('DELETE FROM income WHERE id IN (SELECT income_id FROM invoice WHERE number="' . $this->invoiceNumber . '")');
-        $this->pdo->query('DELETE FROM invoice WHERE number = "'.$this->invoiceNumber.'"');
+        if ($this->invoiceNumber) {
+            $this->pdo->query('DELETE FROM income WHERE id IN (SELECT income_id FROM invoice WHERE number="' . $this->invoiceNumber . '")');
+            $this->pdo->query('DELETE FROM invoice WHERE number = "'.$this->invoiceNumber.'"');
+        }
     }
 
     private function setBusinessData(): void
